Remove More Informative Scan feature flag
