--- client/components/jetpack/fix-all-threats-dialog-new/style.scss
@@ -1,86 +1,0 @@
-.fix-all-threats-dialog-new {
-	&__subheader {
-		font-weight: 600;
-		margin: 8px 0;
-	}
-
-	&__threats-section {
-		display: flex;
-		flex-direction: column;
-		overflow-y: auto;
-	}
-
-	&__card-container {
-		display: flex;
-		flex-wrap: wrap;
-		flex-direction: column;
-		border: 1px solid var( --studio-gray-5 );
-		border-bottom: 0;
-		height: 97px;
-		width: 594px;
-		flex-grow: 1;
-
-		&:last-of-type {
-			border-bottom: 1px solid var( --studio-gray-5 );
-		}
-
-		@include breakpoint-deprecated( '>960px' ) {
-			flex-wrap: nowrap;
-			flex-direction: row;
-			height: 100px;
-		}
-	}
-
-
-
-	&__icon {
-		color: var( --color-primary-40 );
-		flex: 0 0 24px;
-		margin-top: -7px;
-		margin-right: 8px;
-	}
-
-	&__warning {
-		display: flex;
-		margin-top: 16px;
-	}
-
-	&__threats {
-		margin-left: 12px;
-		margin-top: 8px;
-
-		&.is-long-list {
-			&::after {
-				content: '';
-				position: absolute;
-				left: 0;
-				bottom: 100px;
-				width: 100%;
-				height: 75px;
-				background: linear-gradient(
-					0deg,
-					rgba( 255, 255, 255, 1 ) 0%,
-					rgba( 255, 255, 255, 0 ) 100%
-				);
-				pointer-events: none;
-
-				@include breakpoint-deprecated( '>660px' ) {
-					bottom: 70px;
-				}
-			}
-		}
-	}
-
-	&__warning-message {
-		color: var( --color-text-subtle );
-		font-size: $font-body-small;
-	}
-
-	&__btn {
-		width: 100%;
-
-		@include breakpoint-deprecated( '>480px' ) {
-			width: 140px;
-		}
-	}
-}
