--- client/components/jetpack/scan-threats-new/style.scss
@@ -1,152 +1,0 @@
-.scan-threats-new {
-	display: flex;
-
-	&__threats {
-		margin: 0;
-	}
-
-	&__card-header {
-		flex: none;
-		padding: 32px;
-	}
-
-	&__header-message {
-		display: flex;
-		padding-bottom: 8px;
-	}
-
-	&__low-risk {
-		margin-top: 16px;
-	}
-
-	&__threats {
-		.foldable-card {
-			margin-bottom: 0;
-		}
-
-		.foldable-card.card.is-expanded {
-			margin: 0;
-		}
-
-		.foldable-card.is-expanded .foldable-card__content {
-				padding: 0;
-		}
-	}
-
-	&.scan__header {
-		margin-top: 24px;
-		margin-bottom: 16px;
-	}
-
-	&__error {
-		text-align: center;
-		font-size: $font-body-small;
-		line-height: 20px;
-		color: var( --studio-gray-40 );
-		margin-top: 24px;
-	}
-
-	&__rerun {
-		margin-top: 32px;
-		display: flex;
-		flex-direction: column;
-
-		@include breakpoint-deprecated( '>800px' ) {
-			flex-direction: row;
-		}
-
-		.scan-threats-new__rerun-help {
-			font-size: $font-body-small;
-			line-height: 21px;
-			font-style: normal;
-			font-weight: normal;
-			color: var( --studio-gray-60 );
-			text-align: center;
-			letter-spacing: -0.01em;
-			max-width: 576px;
-			height: 42px;
-			margin-left: 72px;
-			margin-right: 72px;
-
-			@include breakpoint-deprecated( '>800px' ) {
-				margin-bottom: 0;
-			}
-		}
-
-		.scan-threats-new__run-scan-main-button {
-			flex: 1 1 40%;
-			padding: 10px 30px;
-
-			@include breakpoint-deprecated( '>800px' ) {
-				flex: 0 0 170px;
-				margin-left: 8px;
-			}
-		}
-	}
-
-	&__run-scan-button.scan-threats-new__run-scan-button {
-		padding: initial;
-		border: initial;
-		font-weight: normal;
-		vertical-align: initial;
-		background: inherit;
-
-		&:hover,
-		&:active,
-		&:focus {
-			background: initial;
-			color: var( --color-link-dark );
-		}
-
-		&:hover,
-		&:active {
-			outline: initial;
-		}
-
-		&:focus {
-			outline: thin dotted;
-			box-shadow: initial;
-			border: initial;
-		}
-	}
-
-	&__buttons {
-		display: flex;
-		border-top: 1px solid var( --studio-gray-5 );
-		padding-top: 32px;
-		flex-direction: space-between;
-		flex-wrap: wrap;
-		align-items: center;
-
-		@include breakpoint-deprecated( '>800px' ) {
-			flex-direction: row;
-		}
-
-		> p {
-			font-size: $font-body;
-			font-weight: normal;
-			// line-height: 24px;
-
-			@include breakpoint-deprecated( '>800px' ) {
-				flex-grow: 1;
-				margin: 0 !important;
-			}
-		}
-	}
-
-	&__fix-all-threats-button {
-		width: 105px;
-		height: 40px;
-		border-radius: 2px;
-		font-weight: 600;
-		@include breakpoint-deprecated( '<800px' ) {
-			flex-grow: 1;
-		}
-	}
-
-	&__options-button {
-		flex: 0 0 50px;
-		padding: 10px 0;
-		margin-left: 16px;
-	}
-}
