--- client/components/jetpack/threat-description-new/style.scss
@@ -1,43 +1,0 @@
-.scan-history-item .threat-description-new,
-.threat-item-new .threat-description-new {
-	overflow: auto;
-	flex: 1;
-	margin-left: 116px;
-	margin-right: 32px;
-
-	@include breakpoint-deprecated( '<960px' ) {
-		padding: 16px;
-		margin: 0;
-	}
-
-	&__section-title {
-		margin-bottom: 20px;
-
-		font-size: $font-body;
-		line-height: 1.3; /* 24/18 */
-	}
-
-	&__section-title-fix {
-		margin-top: 20px;
-		font-size: $font-body;
-	}
-
-	&__section-text {
-		margin-bottom: 16px;
-
-		color: var( --color-text-subtle );
-		font-size: $font-body-small;
-
-		@include breakpoint-deprecated( '>660px' ) {
-			margin-bottom: 24px;
-		}
-	}
-
-	&__alert-filename {
-		font-size: $font-body-small;
-	}
-
-	.marked-lines {
-		font-size: $font-body-small;
-	}
-}
