+++ client/components/jetpack/threat-description/style.scss
@@ -2,16 +2,24 @@
 .threat-item .threat-description {
 	overflow: auto;
 	flex: 1;
+	margin-left: 116px;
+	margin-right: 32px;
+
+	@include breakpoint-deprecated( '<960px' ) {
+		padding: 16px;
+		margin: 0;
+	}
 
 	&__section-title {
 		margin-bottom: 20px;
 
-		font-size: $font-title-small;
+		font-size: $font-body;
 		line-height: 1.3; /* 24/18 */
 	}
 
 	&__section-title-fix {
 		margin-top: 20px;
+		font-size: $font-body;
 	}
 
 	&__section-text {
@@ -23,5 +31,13 @@
 		@include breakpoint-deprecated( '>660px' ) {
 			margin-bottom: 24px;
 		}
+	}
+
+	&__alert-filename {
+		font-size: $font-body-small;
+	}
+
+	.marked-lines {
+		font-size: $font-body-small;
 	}
 }
