+++ client/components/jetpack/threat-item-header/style.scss
@@ -1,4 +1,4 @@
-.threat-item-header-new {
+.threat-item-header {
   &__card-container {
     flex-direction: column;
     margin-bottom: auto;
