+++ client/components/jetpack/threat-item/index.tsx
@@ -5,16 +5,15 @@ import * as React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import ExternalLinkWithTracking from 'calypso/components/external-link/with-tracking';
 import ThreatItemHeader from 'calypso/components/jetpack/threat-item-header';
-import ThreatItemSubheader from 'calypso/components/jetpack/threat-item-subheader';
-import { Threat } from 'calypso/components/jetpack/threat-item/types';
 import { getThreatFix } from 'calypso/components/jetpack/threat-item/utils';
 import { recordTracksEvent } from 'calypso/state/analytics/actions/record';
 import getCurrentRoute from 'calypso/state/selectors/get-current-route';
 import LogItem from '../log-item';
 import ThreatDescription from '../threat-description';
+import type { Threat } from 'calypso/components/jetpack/threat-item/types';
 import type { TranslateResult } from 'i18n-calypso';
-import './style.scss';
 
+import './style.scss';
 interface Props {
 	threat: Threat;
 	isPlaceholder: boolean;
@@ -38,7 +37,6 @@ const ThreatItem: React.FC< Props > = ( {
 	onFixThreat,
 	onIgnoreThreat,
 	isFixing,
-	contactSupportUrl,
 } ) => {
 	const dispatch = useDispatch();
 
@@ -71,7 +69,6 @@ const ThreatItem: React.FC< Props > = ( {
 		[ isFixing, onFixThreat, threat ]
 	);
 
-	/* eslint-disable wpcalypso/jsx-classname-namespace */
 	const getFix = React.useCallback( (): TranslateResult | undefined => {
 		if ( threat.status === 'fixed' ) {
 			return;
@@ -81,7 +78,7 @@ const ThreatItem: React.FC< Props > = ( {
 			return (
 				<>
 					{ ! threat.rows && (
-						<p className="threat-description__section-text">
+						<p className="threat-item threat-description__section-text">
 							{ translate(
 								'Jetpack Scan cannot automatically fix this threat. We suggest that you resolve the threat manually: ' +
 									'ensure that WordPress, your theme, and all of your plugins are up to date, and remove ' +
@@ -90,7 +87,7 @@ const ThreatItem: React.FC< Props > = ( {
 						</p>
 					) }
 					{ threat.rows && (
-						<p className="threat-description__section-text">
+						<p className="threat-item threat-description__section-text">
 							{ translate(
 								'Jetpack Scan cannot automatically fix this threat. We suggest that you resolve the threat manually: ' +
 									'ensure that WordPress, your theme, and all of your plugins are up to date, and remove or edit ' +
@@ -99,7 +96,7 @@ const ThreatItem: React.FC< Props > = ( {
 						</p>
 					) }
 					{ 'current' === threat.status && (
-						<p className="threat-description__section-text">
+						<p className="threat-item threat-description__section-text">
 							{ translate(
 								'If you need more help to resolve this threat, we recommend {{strong}}Codeable{{/strong}}, a trusted freelancer marketplace of highly vetted WordPress experts. ' +
 									'They have identified a select group of security experts to help with these projects. ' +
@@ -116,9 +113,17 @@ const ThreatItem: React.FC< Props > = ( {
 			);
 		}
 
-		return <p className="threat-description__section-text">{ getThreatFix( threat.fixable ) }</p>;
-	}, [ contactSupportUrl, threat ] );
-	/* eslint-enable wpcalypso/jsx-classname-namespace */
+		return (
+			<p className="threat-item threat-description__section-text">
+				{ getThreatFix( threat.fixable ) }
+				<p>
+					{ translate(
+						'Jetpack Scan is able to automatically fix this threat for you. Since it will replace the affected file or directory the site’s look-and-feel or features can be compromised. We recommend that you check if your latest backup was performed successfully in case a restore is needed.'
+					) }
+				</p>
+			</p>
+		);
+	}, [ threat ] );
 
 	const isFixable = React.useMemo(
 		() => threat.fixable && ( threat.status === 'current' || threat.status === 'ignored' ),
@@ -156,8 +161,6 @@ const ThreatItem: React.FC< Props > = ( {
 				'is-current': threat.status === 'current',
 			} ) }
 			header={ <ThreatItemHeader threat={ threat } isStyled={ true } /> }
-			subheader={ <ThreatItemSubheader threat={ threat } isFixable={ isFixable } /> }
-			{ ...( threat.status === 'current' ? { highlight: 'error' } : {} ) }
 			clickableHeader={ true }
 			onClick={ onOpenTrackEvent }
 		>
