+++ config/jetpack-cloud-development.json
@@ -37,7 +37,6 @@
 		"dev/preferences-helper": true,
 		"jetpack-cloud": true,
 		"jetpack/activity-log-sharing": true,
-		"jetpack/more-informative-scan": true,
 		"jetpack/partner-portal-payment": true,
 		"jetpack/pricing-page": true,
 		"jetpack/pricing-page-annual-only": true,
