+++ config/jetpack-cloud-horizon.json
@@ -31,7 +31,6 @@
 		"current-site/stale-cart-notice": false,
 		"checkout/google-pay": true,
 		"jetpack-cloud": true,
-		"jetpack/more-informative-scan": true,
 		"jetpack/search-product": true,
 		"jetpack/pricing-page": true,
 		"jetpack/pricing-page-annual-only": true,
