+++ config/jetpack-cloud-stage.json
@@ -34,7 +34,6 @@
 		"checkout/google-pay": false,
 		"jetpack-cloud": true,
 		"jetpack/activity-log-sharing": true,
-		"jetpack/more-informative-scan": true,
 		"jetpack/partner-portal-payment": true,
 		"jetpack/pricing-page": true,
 		"jetpack/pricing-page-annual-only": true,
